+++ packages/stream_feeds/CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- [BREAKING] Renamed `AppLifecycleStateProvider` to `LifecycleStateProvider` and `AppLifecycleState` to `LifecycleState`.
+
 ## 0.2.0
 - [BREAKING] Update API client code, specifically the FeedOwnCapability enum.
 - Fix unknown enums for `List<FeedOwnCapability>` in `GetOrCreateFeedResponse` to be `FeedOwnCapability.unknown`.